requirements.txt
appdirs==1.4.4
asttokens==3.0.0
bonobo==0.6.4
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
decorator==5.2.1
docopt==0.6.2
duckdb==1.2.0
exceptiongroup==1.2.2
executing==2.2.0
fs==2.4.16
graphviz==0.8.4
icecream==2.1.4
idna==3.10
ipython==8.18.1
jedi==0.19.2
Jinja2==2.11.3
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mondrian==0.8.1
mysql-connector==2.2.9
packaging==19.2
parso==0.8.4
pbr==6.1.0
pexpect==4.9.0
prompt_toolkit==3.0.50
psutil==5.9.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.1
python-slugify==1.2.6
requests==2.32.3
six==1.17.0
stack-data==0.6.3
stevedore==1.32.0
traitlets==5.14.3
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.3.0
wcwidth==0.2.13
whistle==1.0.1
